backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-redis>=5.4
celery>=5.3
redis>=5.0
psycopg[binary]>=3
requests>=2.31
python-dateutil>=2.9
pydantic>=2.8
trafilatura>=1.7
uvicorn>=0.30
gunicorn>=21.2
google-api-python-client>=2.137
google-auth>=2.35
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2
tldextract>=5.1
feedparser>=6.0.11
